+++ test/isolate_rpc_client_test.dart
@@ -1,6 +1,6 @@
 import 'dart:isolate';
 
-import 'package:flutter_litert/flutter_litert.dart';
+import 'package:flutter_litert/native.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 /// Isolate that serves via [serveIsolateRpc] (which acks dispose).