+++ test/isolate_rpc_server_test.dart
@@ -1,6 +1,6 @@
 import 'dart:isolate';
 
-import 'package:flutter_litert/flutter_litert.dart';
+import 'package:flutter_litert/native.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 /// Isolate entry that serves a few ops via [serveIsolateRpc].