+++ test/native/interpreter_options_fallback_test.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_litert/flutter_litert.dart';
+import 'package:flutter_litert/native.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {