+++ test/native/interpreter_pool_test.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_litert/flutter_litert.dart';
+import 'package:flutter_litert/native.dart';
 
 File get _modelFile =>
     File('${Directory.current.path}/test/assets/training_model.tflite');