+++ test/util/tensor_views_test.dart
@@ -3,7 +3,7 @@ import 'dart:math' as math;
 import 'dart:typed_data';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_litert/flutter_litert.dart';
+import 'package:flutter_litert/native.dart';
 
 File get _modelFile =>
     File('${Directory.current.path}/test/assets/training_model.tflite');